keeping smaller value in files mtime
As my wife Silvia made me notice, the oldest mtime is the most reliable,
as the copies having a wrong mtime, will most likely have a 'future' value
w.r.t. the original one.

Of course this assumes that the clock is correct on the device that made
the copy, but in modern times this is a very reasonable assumption!